ファイル削除の部分を streamFileManager へ分割
